Reworked logic to create DataFrame to cache vals
Reworked logic to create a one off DataFrame of Kmer, position, p-value for each position in the alignment (and for each sequence with an alignment at that position).

Now HMP/windows can be generated much quicker, since p-values for each window do not need to be recalculated for each window_size.

For each window, just takes 'slice' of dataframe positions and takes HMP of p-vals of said rows.

DataFrame itself is very memory intensive. May need to create SQL db for querying or array.